Move store key utils to framework
